% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumption_checking.R
\name{cor_graphic}
\alias{cor_graphic}
\title{Correlation Plots}
\usage{
cor_graphic(data)
}
\arguments{
\item{data}{A numeric vector, matrix or data frame to be used as input to
\code{cor}}
}
\description{
\code{cor_graphic} uses the "corrplot" package to create a nice graphic with a
color-coded correlation matrix and correlation plot side-by-side
}
