% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumption_checking.R
\name{jcreg_boxplot}
\alias{jcreg_boxplot}
\title{Boxplot of residuals}
\usage{
jcreg_boxplot(model)
}
\arguments{
\item{model}{A linear regression model of class \code{stats::lm}}
}
\description{
Uses ggplot2::ggplot to create a nicely formatted box plot of the residuals of a model
}
\examples{
jcreg_boxplot(lm(mpg ~ cyl, data = mtcars))

}
