% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumption_checking.R
\name{jcreg_dffits}
\alias{jcreg_dffits}
\title{DFFITS Plot}
\usage{
jcreg_dffits(model, nLabels = 3)
}
\arguments{
\item{model}{A linear regression model of class \code{stats::lm}}

\item{nLabels}{The number of potential influential points to label. Defaults
to 3. A value of 0 means do not label potential influential points.}
}
\description{
Plots the DFFITS against observation number.
}
