% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumption_checking.R
\name{point_matrix}
\alias{point_matrix}
\title{Scatterplot matrix}
\usage{
point_matrix(data, pch = 19)
}
\arguments{
\item{data}{The coordinates of points given as numeric columns of a matrix or
data frame. See parameter \code{x} in \code{pairs()} for more detail.}

\item{pch}{Point type passed on to \code{pairs}.}
}
\description{
\code{point_matrix} returns a basic scatterplot matrix of the data input.
}
\details{
This function uses pairs(data) to generate a scatterplot matrix with square
plots and solid points. Note that for best results you should set your plot
size to be square.
}
\examples{
point_matrix(bodyfat)
point_matrix(mtcars[ ,c("mpg", "disp", "hp", "drat")])

}
