% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumption_checking.R
\name{print.summary.glm.jcreg}
\alias{print.summary.glm.jcreg}
\title{Summarizing Generalized Linear Model Fits}
\usage{
\method{print}{summary.glm.jcreg}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\description{
This function is a slilght modification of stats::print.summary.glm
}
