% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumption_checking.R
\name{print.var_selection}
\alias{print.var_selection}
\title{Print \code{var_selection} Object}
\usage{
\method{print}{var_selection}(obj)
}
\arguments{
\item{obj}{An object of class \code{var_selection}}
}
\description{
Implements a print function for an object of type \code{var_selection}
}
