% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumption_checking.R
\name{rpred_col}
\alias{rpred_col}
\title{Residuals vs Predictor Plot}
\usage{
rpred_col(data, residuals, predictor)
}
\arguments{
\item{data}{A dataframe containing one or more columns including the
predictor variable.}

\item{residuals}{A vector or a dataframe containing the residuals assosiated
with the data. Its length must be the same as the numer of rows in \code{data}.}

\item{predictor}{A string. The name of the column in \code{data} to be used as the
predictor variable.}
}
\description{
A helper function for resid_vs_pred. Plots the residuals vs the predictor for
a single predictor variable. Returns a ggplot2::ggplot object.
}
