% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumption_checking.R
\name{summary.glm}
\alias{summary.glm}
\title{Summarizing Generalized Linear Model Fits}
\usage{
\method{summary}{glm}(
  object,
  dispersion = NULL,
  correlation = FALSE,
  symbolic.cor = FALSE,
  ...
)
}
\description{
This function is identical to stats::summary.glm except the class of the
output is changed for more versitile output
}
