% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumption_checking.R
\name{jcreg_av}
\alias{jcreg_av}
\title{Added-Variable Plots}
\usage{
jcreg_av(model)
}
\arguments{
\item{model}{A linear regression model of class \code{stats::lm}}
}
\value{
Returns the output of car::avPlots
}
\description{
Wrapper for \code{car::avPlots} with square plots and well laid out.
}
\examples{
jcreg_av(lm(brozek ~ ., data = bodyfat))

}
