% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumption_checking.R
\name{resid_vs_fitted}
\alias{resid_vs_fitted}
\title{Residuals Vs Fitted Values Plot}
\usage{
resid_vs_fitted(model)
}
\description{
\code{resid_vs_fitted} uses \code{autoplot} to create a square residuals vs fitted
values plot with a nice theme
}
\details{
@param model A linear regression model of class \code{stats::lm}

@export
}
