% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumption_checking.R
\name{resid_vs_fitted}
\alias{resid_vs_fitted}
\title{Residuals Vs Fitted Values Plot}
\usage{
resid_vs_fitted(model)
}
\arguments{
\item{model}{A linear regression model of class \code{stats::lm}.}
}
\value{
Returns a ggplot object
}
\description{
\code{resid_vs_fitted} uses \code{autoplot} from \code{ggfortify} to create a square
residuals vs fitted values plot with a nice theme
}
\examples{
resid_vs_fitted(lm(brozek ~ ., data = bodyfat))

}
