% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumption_checking.R
\name{resid_vs_pred}
\alias{resid_vs_pred}
\title{Matrix of Residuals Vs Predictor Plots}
\usage{
resid_vs_pred(model)
}
\arguments{
\item{A}{linear regression model of class \code{stats::lm}.}
}
\value{
Returns a gridExtra object with class "gtable", "gTree", "grob", and
"gDesc".
}
\description{
\code{resid_vs_pred} uses ggplot2::ggplot to create a matrix of residual vs
predictor plots
}
\examples{
resid_vs_pred(lm(brozek ~ ., data = bodyfat))

}
